changed test for messageParsingService
